+++ index.html
@@ -7,25 +7,10 @@
     <link rel="stylesheet" href="./index.css" />
 </head>
 <body>
-    <!-- <div id="header">
-        <h1>Rachna where are you</h1> -->
-
-    <!-- </div> --> 
-    
     <div id="root">
         <h1>i m here</h1>
         <h1>I am here and you</h1>
     </div>
-    
-
-    <script type="module" src="./App.js"></script>
-
-   <!-- <script>
-    const heading=document.createElement("h1");
-    heading.innerHTML= "Hello world from javascript";
-    const root=document.getElementById("root");
-    root.appendChild(heading)
-   </script> -->
-    
+    <script type="module" src="./src/App.js"></script>
 </body>
 </html>
