+++ src/components/RestaurantCard.js
@@ -1,0 +1,32 @@
+import { CDN_URL } from "../utils/constants";
+
+const RestaurantCard=(props)=>{
+    const { resData } = props;
+
+   const {
+    cloudinaryImageId,
+    name,
+    avgRating,
+    cuisines,
+    costForTwo,
+    deliveryTime,
+  } = resData?.data;
+
+    return(     
+        <div className="resturant-card">
+            <div>
+                <img alt='res-logo' src= {
+                          CDN_URL+cloudinaryImageId
+        }/>
+                <h3>{name}</h3>
+                <h4>{cuisines.join(",")}</h4>
+                <h4>{avgRating} stars</h4>
+                <h4>₹{costForTwo / 100} FOR TWO</h4>
+                <h4>{deliveryTime} minutes</h4>
+            </div>
+        
+        </div>
+    )
+};
+
+export default RestaurantCard;
